requirements.txt
git+https://github.com/Rapptz/discord.py
discord.py==1.6.0
datetime
pendulum